+++ catroidTest/src/org/catrobat/catroid/uitest/ui/dialog/StageDialogTest.java
@@ -70,7 +70,9 @@ public class StageDialogTest extends BaseActivityInstrumentationTestCase<MainMen
 	protected void tearDown() throws Exception {
 		// normally super.teardown should be called last
 		// but tests crashed with Nullpointer
-		ProjectManager.getInstance().deleteCurrentProject();
+		if(StorageHandler.getInstance().projectExists(ProjectManager.getInstance().getCurrentProject().getName())) {
+			ProjectManager.getInstance().deleteCurrentProject();
+		}
 		super.tearDown();
 	}
